requirements.txt
requests~=2.27.1
opencv-python~=4.5.5.62
cvzone~=1.5.6

matplotlib~=3.5.2
setuptools~=57.0.0
certifi~=2021.10.8
pymongo~=4.1.1
python-dotenv~=0.20.0